+++ packages/firebase/src/lib/common/auth/oidc/index.ts
@@ -1,2 +1,5 @@
 export * from './oidc.error';
+export * from './oidc.base';
+export * from './oidc.interaction';
+export * from './oidc.profile';
 export * from './oidc';