+++ packages/firebase/src/lib/model/oidcmodel/index.ts
@@ -4,6 +4,3 @@ export * from './oidcmodel.query';
 export * from './oidcmodel.api';
 export * from './oidcmodel.action';
 export * from './oidcmodel.data';
-export * from './oidcmodel.interaction';
-export * from './oidcmodel.interaction.oauth';
-export * from './oidc.profile';